requirement.txt
flask==2.3.3
firebase-admin==6.2.0
qrcode[pil]==7.4.2
pillow==10.0.0
python-dotenv==1.0.0
pyrebase4==4.6.0
requests==2.31.0
cloudinary==1.36.0